% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScoreStats.R
\name{calcScoreStats}
\alias{calcScoreStats}
\title{calcScoreStats.R}
\usage{
calcScoreStats(x, UgTerm, ggTerm, gtTermCommon, gtHalfL, gtHalfR, solveItt)
}
\arguments{
\item{x}{n*1 vector of genotypes.}

\item{UgTerm}{n*1 vector multiplier for the score statistic.}

\item{ggTerm}{n*1 vector multiplier for the Igg term of the variance.}

\item{gtTermCommon}{n*p matrix multiplier for the common part of the Igt term of the variance.}

\item{gtHalfL}{n*(nknots+1) matrix multiplier for one half of the unique part of the Igt term of the variance.}

\item{gtHalfR}{n*(nknots+1) matrix multiplier for one half of the unique part of the Igt term of the variance.}

\item{solveItt}{p*p inverse of the Itt matrix from ICSKAT_fit_null().}
}
\value{
A 2*1 vector with the test statistic and then p-value.
}
\description{
Function that is applied in ICsingleSNP() to calculate a score statistic and p-value for each
column of an n*p genotype matrix.
}
