% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSqMatchFast.R
\name{chiSqMatchFast}
\alias{chiSqMatchFast}
\title{chiSqMatchFast.R}
\usage{
chiSqMatchFast(lambdaVec)
}
\arguments{
\item{lambdaVec}{Numeric vector holding the eigenvalues of the "middle" term.}
}
\value{
A list with the elements:
\item{sigmaQrho}{Standard deviation of the mixture distribution}
\item{muQrho}{Mean of the mixture distribution}
\item{delta}{Noncentrality parameter of the matched distribution}
\item{l}{Degrees of freedom of the matched distribution}
}
\description{
Match the moments of a mixture of scaled chi-square random variables to a single non-central chi-square.
For the quadratic form case where the mean of the multivariate normal is 0.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
ICskat(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt,
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))
Rrho <- matrix(data=tempRho, nrow=10, ncol=10)
diag(Rrho) <- 1
toDecomp <- Rrho \%*\% icskatOut$sig_mat
Aeigen <- eigen(toDecomp, symmetric = TRUE, only.values = TRUE)
tempLambda <- Aeigen$values
chiSqMatchFast(lambdaVec = tempLambda) 

}
