% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_geno.R
\name{clean_geno}
\alias{clean_geno}
\title{Clean gds genotype function, called by geno_sjlife_gds.R}
\usage{
clean_geno(n, geno_data, map_file)
}
\arguments{
\item{n}{only keep SNPs with at least n non-missing entries}

\item{geno_data}{genotype matrix from the getGenotype() function}
}
\value{
An n*m matrix where m depends on the number of SNPs that passed QC.
}
\description{
Clean gds genotype function, called by geno_sjlife_gds.R
}
