% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSNPalt.R
\name{coxphFn}
\alias{coxphFn}
\title{coxphFn}
\usage{
coxphFn(x, xMat, midTime, midEvent, p)
}
\arguments{
\item{x}{n*1 genotype vector.}

\item{xMat}{n*p matrix of non-genotype covariates.}

\item{midTime}{n*1 vector of event times imputed to be right-censored times using the midpoint imputation method.}

\item{midEvent}{n*1 vector event indicators (0 for censoring, 1 for event) after times have been transformed to right-censored observations.}

\item{p}{scalar, number of columns in xMat.}
}
\value{
A scalar p-value for testing the effect of the genotype in survreg() Weibull model.
}
\description{
Function to be applied over gMat to get p-values from coxPH().
}
