% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_IC_data.r
\name{createInt}
\alias{createInt}
\title{Called by gen_IC_data() to turn the actual outcome times and observation times into interval-censored
outcomes for each subject. Apply this with mapply over a data.frame of visit times, pass in the exact times.
Returns 999 instead of Inf.}
\usage{
createInt(obsTimes, eventTime)
}
\arguments{
\item{obsTimes}{A vector of all the times a subject is observed.}

\item{eventTime}{The exact event time for the subject.}
}
\value{
A 2*1 vector which is the interval of the event time
}
\description{
Called by gen_IC_data() to turn the actual outcome times and observation times into interval-censored
outcomes for each subject. Apply this with mapply over a data.frame of visit times, pass in the exact times.
Returns 999 instead of Inf.
}
\examples{
obsTimes <- 1:10
eventTime <- 7.7
createInt(obsTimes, eventTime)

}
