% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fIntegrate.R
\name{fIntegrate}
\alias{fIntegrate}
\title{fIntegrate.R}
\usage{
fIntegrate(
  x,
  muK1,
  sigmaK1,
  sigmaZeta,
  kappaLambda,
  QrhoDF,
  kurtK1 = NULL,
  matchKurt = FALSE
)
}
\arguments{
\item{x}{Variable to be integrated over, can be a vector or scalar.}

\item{muK1}{Mean of the mixture of chi-squares that are the first part of the kappa variable.
When we do bootstrap we often pass in the mean of the entire kappa, since the mean of zeta is supposed to be 0.}

\item{sigmaK1}{Standard deviation of the entire kappa term.}

\item{sigmaZeta}{Standard deviation of the zeta part of the kappa variable.}

\item{kappaLambda}{Eigenvalues that weight the mixture of chi-squares that are the first part of the kappa variable.}

\item{QrhoDF}{The data frame output from calling QrhoIC().}
}
\value{
The value of the integrand at x.
}
\description{
The integrand in the SKATO p-value, pass it to a numerical integration function like integrate(), uses
Davies method instead of Liu to calculate the probability in the integrand.
}
