% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fIntegrateLiu.R
\name{fIntegrateLiu}
\alias{fIntegrateLiu}
\title{fIntegrateLiu.R}
\usage{
fIntegrateLiu(x, muK1, sigmaK1, sigmaZeta, kappaLambda, QrhoDF, dfK1)
}
\arguments{
\item{x}{Variable to be integrated over, can be a vector or scalar.}

\item{muK1}{Mean of the mixture of chi-squares that are the first part of the kappa variable.}

\item{sigmaK1}{Standard deviation of the mixture of chi-squares that are the first part of the kappa variable.}

\item{sigmaZeta}{Standard deviation of the zeta part of the kappa variable.}

\item{kappaLambda}{Eigenvalues that weight the mixture of chi-squares that are the first part of the kappa variable.}

\item{QrhoDF}{The data frame output from calling QrhoIC().}

\item{dfK1}{The degrees of freedom from the approximated chi-square.}
}
\value{
The value of the integrand at x.
}
\description{
The integrand in the SKATO p-value when using Liu instead of Davies method, pass it to a numerical integration function like integrate().
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
myoutput <- ICskat(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt,
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))
QrhoIC(rhoVec = seq(from=0, to=1, by=0.1), icskatOut = myoutput)
fIntegrateLiu(x=1, muK1=10, sigmaK1=2, sigmaZeta=2, kappaLambda = rep(1, 10), QrhoDF=QrhoDF, dfK1 = 10)

}
