% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_IC_data.r
\name{gen_IC_data}
\alias{gen_IC_data}
\title{gen_IC_data.R}
\usage{
gen_IC_data(bhFunInv, obsTimes, windowHalf, etaVec, probMiss = 0.1)
}
\arguments{
\item{bhFunInv}{A function, the inverse of the baseline hazard function.}

\item{obsTimes}{Vector of the intended observation times.}

\item{windowHalf}{The amount of time before or after the intended obsTimes that a visit might take place.}

\item{etaVec}{n*1 linear predictor in either the proportional odds or proportional hzards model.}

\item{probMiss}{The probability of missing any given visit.}
}
\value{
A list with the elements:
\item{deltaVec}{n*1 indicator of whether the event has happened for each subject}
\item{tVec}{n*1 vector of exact event times}
\item{leftTimes}{n*1 vector of left side of interval for each subject}
\item{rightTimes}{n*1 vector of right side of interval for each subject}
}
\description{
Generate interval-censored data under the proportional odds/PH model given a baseline hazard function and
some information about observation times.
}
\examples{
bhFunInv <- function(x) {sqrt(2*x)}
obsTimes <- seq(from=0.5, to=3.5, by=0.5)
 xMat <- cbind(rnorm(n), rbinom(n=n, size=1, prob=0.5))
betaVec <- c(-0.5, 0.5)
etaVec <- xMat \%*\% betaVec
outcomeDat <- gen_IC_data(bhFunInv = bhFunInv, obsTimes = obsTimes, windowHalf = 0.25, etaVec = etaVec, n=n)

}
