% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICsingleSNP.R
\name{ICsingleSNP}
\alias{ICsingleSNP}
\title{ICsingleSNP.R}
\usage{
ICsingleSNP(left_dmat, right_dmat, lt, rt, obs_ind, tpos_ind, gMat,
  null_beta, solveItt, p)
}
\arguments{
\item{left_dmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{right_dmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{lt}{n*1 vector of left side of time interval.}

\item{rt}{n*1 vector of right side of time interval.}

\item{obs_ind}{n*1 vector of whether the event was observed before last follow-up.}

\item{tpos_ind}{n*1 vector of whether the event was observed after follow-up started (t>0).}

\item{gMat}{n*p genotype matrix.}

\item{null_beta}{(p+nknots+2)*1 vector of coefficients for null model.}

\item{Itt}{(p+nknots+2)*(p+nknots+2) Fisher information matrix for null model coefficients.}
}
\value{
A list with the elements:
\item{testStatsVec}{p*1 vector of score test statistics}
\item{pVec}{p*1 vector of score test p-values}
}
\description{
Same as ICSKAT() except do a separate score test for each SNP in gMat, one at a time.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
ICsingleSNP(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt, 
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))

}
