% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSKAT.R
\name{ICskat}
\alias{ICskat}
\title{ICSKAT.R}
\usage{
ICskat(left_dmat, right_dmat, lt, rt, obs_ind, tpos_ind, gMat, null_beta,
  Itt)
}
\arguments{
\item{left_dmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{right_dmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{obs_ind}{n*1 vector of whether the event was observed before last follow-up.}

\item{tpos_ind}{n*1 vector of whether the event was observed after follow-up started (t>0).}

\item{gMat}{n*q genotype matrix.}

\item{null_beta}{(p+nknots+2)*1 vector of coefficients for null model.}

\item{Itt}{(p+nknots+2)*(p+nknots+2) Fisher information matrix for null model coefficients.}
}
\value{
A list with the elements:
\item{p_SKAT}{ICSKAT p-value}
\item{p_burden}{IC burden test p-value}
\item{complex}{Indicator of whether the SKAT variance matrix was positive definite}
\item{sig_mat}{The covariance matrix of the score equations for genetic effects when treated as fixed effects} 
\item{skatQ}{SKAT test statistic}
\item{burdenQ}{Burden test statistic}
}
\description{
Calculate the test statistic and p-value for interval-censored skat.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
ICskat(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt, 
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))

}
