% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSKAT_fit_null.R
\name{ICSKAT_fit_null}
\alias{ICSKAT_fit_null}
\title{ICSKAT_fit_null.R}
\usage{
ICSKAT_fit_null(init_beta, left_dmat, obs_ind, tpos_ind, right_dmat, lt,
  rt, checkpoint = FALSE, eps = 10^(-6))
}
\arguments{
\item{init_beta}{(p+nknots+2)*1 vector of coefficients to initialize the Newton-Raphson.}

\item{left_dmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{obs_ind}{n*1 vector of whether the event was observed before last follow-up.}

\item{tpos_ind}{n*1 vector of whether the event was observed after follow-up started (t>0).}

\item{right_dmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{checkpoint}{Boolean tells the function to print when each iteration completes.}

\item{eps}{Stop when the L2 norm of the difference in model coefficients reaches this limit.}
}
\value{
A list with the elements:
\item{beta_fit}{(p+nknots+2)*1 vector of fitted coefficients under null model.}
\item{iter}{Number of iterations needed to converge.}
\item{Itt}{Fisher information matrix for the fitted coefficients.}
}
\description{
Fit the null model (cubic basis spline for baseline cumulative hazard and coefficients
for non-genetic coefficiens) for interval-censored skat.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
ICSKAT_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat, lt=lt, rt=rt,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))

}
