% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSKATO.R
\name{ICSKATO}
\alias{ICSKATO}
\title{ICSKATO.R}
\usage{
ICSKATO(icskatOut, liu = TRUE, rhoVec = c(0, 0.01, 0.04, 0.09, 0.25,
  0.5, 1))
}
\arguments{
\item{icskatOut}{The output list from ICSKAT().}

\item{liu}{Boolean for whether to use Liu moment matching approximation for p-value or Davies method.}

\item{rhoVec}{Vector of rhos to search over.}
}
\value{
A list with the elements:
\item{pval}{SKATO p-value}
\item{QrhoDF}{Data frame containing the distribution and p-value for each Krho.}
\item{r}{The rank of the cholesky decomposition of the kappa part}
\item{intDavies}{Boolean denoting whether integration was with Davies (true) or Liu method (false)}
\item{err}{Boolean, did the integration run into an error}
}
\description{
Calculate SKATO test for ICSKAT.
}
