% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_IC_dmat.R
\name{make_IC_dmat}
\alias{make_IC_dmat}
\title{make_IC_dmat.R}
\usage{
make_IC_dmat(xMat, lt, rt, nknots = 1)
}
\arguments{
\item{xMat}{n*p matrix of non-genetic covariates.}

\item{lt}{n*1 vector with left end of intervals (min is 0).}

\item{rt}{n*1 vector with right end of intervals.}

\item{nknots}{number of knots to use for cubic spline basis (default is 1).}
}
\value{
A list with the elements:
\item{right_dmat}{n*(p+nknots+2) design matrix for right end of interval.}
\item{left_dmat}{n*(p+nknots+2) design matrix for left end of interval.}
#'
}
\description{
Puts together the entire design matrix for both the left and right ends of the
interval, pasting together the non-genetic covariates with the cubic spline basis.
}
\examples{
xMat <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
make_IC_dmat(xMat=xMat, lt=lt, rt=rt)

}
