% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_visit.R
\name{matchVisit}
\alias{matchVisit}
\title{match_visit.R}
\usage{
matchVisit(draw, visitTimes)
}
\arguments{
\item{draw}{(s+1)-length vector of all 0s except for one 1, which is the failure interval.}

\item{visitTimes}{s-length vector where is the number of inspection times for a subject.}
}
\value{
n*(s+1) matrix where element (i,j) holds the probability that subject i will fail in interval j.
}
\description{
Match visit to a time for model-based bootstrap with interval-censored data.
}
