% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QrhoIC.R
\name{QrhoIC}
\alias{QrhoIC}
\title{QrhoIC.R}
\usage{
QrhoIC(rhoVec, icskatOut, liu = TRUE)
}
\arguments{
\item{rhoVec}{Numeric vector of the rho values to use in SKATO.}

\item{null_beta}{The output (a list) from a call to the ICSKAT() function.}

\item{Itt}{Boolean for whether to use Liu (TRUE) or Davies (FALSE) method in calculating p-values.}
}
\value{
Data frame holding the pvalue + test statistic for each fixed rho, the matched noncentrality + degree of freedom parameters
according to the Liu method, and the mean and variance of each Qrho.
}
\description{
Calculate the test statistic, distribution, and p-value for each value of Krho in SKATO.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
myoutput <- ICskat(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt,
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))
QrhoIC(rhoVec = seq(from=0, to=1, by=0.1), icskatOut = myoutput)

}
