% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSNPalt.R
\name{survregFn}
\alias{survregFn}
\title{survregFn}
\usage{
survregFn(x, xMat, leftTime2, rightTime2, p)
}
\arguments{
\item{x}{n*1 genotype vector.}

\item{xMat}{n*p matrix of non-genotype covariates.}

\item{leftTime2}{n*1 vector of left interval times in the format of Surv() interval2 type, i.e NA for left or right censored observations.}

\item{rightTime2}{n*1 vector of right interval times in the format of Surv() interval2 type, i.e NA for left or right censored observations.}

\item{p}{scalar, number of columns in xMat.}
}
\value{
A scalar p-value for testing the effect of the genotype in survreg() Weibull model.
}
\description{
Function to be applied over gMat to get p-values from survreg().
}
