% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScoreStats.R
\name{calcScoreStats}
\alias{calcScoreStats}
\title{calcScoreStats.R}
\usage{
calcScoreStats(x, UgTerm, ggTerm, gtTermCommon, gtHalfL, gtHalfR, solveItt)
}
\arguments{
\item{x}{n*1 vector of genotypes.}

\item{UgTerm}{n*1 vector multiplier for the score statistic.}

\item{ggTerm}{n*1 vector multiplier for the Igg term of the variance.}

\item{gtTermCommon}{n*p matrix multiplier for the common part of the Igt term of the variance.}

\item{gtHalfL}{n*(nknots+1) matrix multiplier for one half of the unique part of the Igt term of the variance.}

\item{gtHalfR}{n*(nknots+1) matrix multiplier for one half of the unique part of the Igt term of the variance.}

\item{solveItt}{p*p inverse of the Itt matrix from ICSKAT_fit_null().}
}
\value{
A 2*1 vector with the test statistic and then p-value.
}
\description{
Function that is applied in ICsingleSNP() to calculate a score statistic and p-value for each
column of an n*p genotype matrix.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
ICskat(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt, 
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))

}
