% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fIntegrateLiu.R
\name{fIntegrateLiu}
\alias{fIntegrateLiu}
\title{fIntegrateLiu.R}
\usage{
fIntegrateLiu(x, muK1, sigmaK1, QrhoDF, dfK1)
}
\arguments{
\item{x}{Variable to be integrated over, can be a vector or scalar.}

\item{muK1}{Mean of the mixture of chi-squares that are the first part of the kappa variable.}

\item{sigmaK1}{Standard deviation of the mixture of chi-squares that are the first part of the kappa variable.}

\item{QrhoDF}{The data frame output from calling QrhoIC().}

\item{dfK1}{The degrees of freedom from the approximated chi-square.}
}
\value{
The value of the integrand at x.
}
\description{
The integrand in the SKATO p-value when using Liu instead of Davies method, pass it to a numerical integration function like integrate().
}
