% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICsingleSNP.R
\name{ICsingleSNP}
\alias{ICsingleSNP}
\title{ICsingleSNP.R}
\usage{
ICsingleSNP(
  left_dmat,
  right_dmat,
  lt,
  rt,
  obs_ind,
  tpos_ind,
  gMat,
  null_beta,
  solveItt,
  p
)
}
\arguments{
\item{left_dmat}{n*(p+nknots+2) design matrix for left end of interval.}

\item{right_dmat}{n*(p+nknots+2) design matrix for right end of interval.}

\item{lt}{n*1 vector of left side of time interval.}

\item{rt}{n*1 vector of right side of time interval.}

\item{obs_ind}{n*1 vector of whether the event was observed before last follow-up.}

\item{tpos_ind}{n*1 vector of whether the event was observed after follow-up started (t>0).}

\item{gMat}{n*q genotype matrix.}

\item{null_beta}{(p+nknots+2)*1 vector of coefficients for null model.}

\item{solveItt}{Inverse of (p+nknots+2)*(p+nknots+2) Fisher information matrix for null model coefficients.}

\item{p}{number of non-SNP covariates.}
}
\value{
A list with the elements:
\item{testStatsVec}{p*1 vector of score test statistics}
\item{pVec}{p*1 vector of score test p-values}
}
\description{
Burden test from ICSKAT() except do a separate burden test for each SNP in gMat, one at a time.
}
\examples{
gMat <- matrix(data=rbinom(n=200, size=2, prob=0.3), nrow=100)
xMat <- matrix(data=rnorm(200), nrow=100)
bhFunInv <- function(x) {x}
obsTimes <- 1:5
etaVec <- rep(0, 100)
outcomeDat <- gen_IC_data(bhFunInv = bhFunInv, obsTime = obsTime, windowHalf = 0.1,
probMiss = 0.1, etaVec = etaVec)
lt <- outcomeDat$leftTimes
rt <- outcomeDat$rightTimes
tpos_ind <- as.numeric(lt > 0)
obs_ind <- as.numeric(rt != Inf)
dmats <- make_IC_dmat(xMat, lt, rt, obs_ind, tpos_ind)
nullFit <- ICSKAT_fit_null(init_beta = rep(0, 5), left_dmat = dmats$left_dmat, right_dmat=dmats$right_dmat,
obs_ind = obs_ind, tpos_ind = tpos_ind, lt = lt, rt = rt)
solveItt  <- solve(nullFit$Itt)
ICsingleSNP(left_dmat = dmats$left_dmat, right_dmat=dmats$right_dmat, lt = lt, rt = rt,
obs_ind = obs_ind, tpos_ind = tpos_ind, gMat = gMat, null_beta = nullFit$beta_fit, solveItt = solveItt, p=2)

}
