% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSNPalt.R
\name{singleSNPalt}
\alias{singleSNPalt}
\title{singleSNPalt.R}
\usage{
singleSNPalt(lt, rt, tpos_ind, obs_ind, xMat, gMat, coxph = TRUE,
  survreg = TRUE)
}
\arguments{
\item{lt}{n*1 vector of left side of time interval.}

\item{rt}{n*1 vector of right side of time interval.}

\item{tpos_ind}{n*1 vector of whether the event was observed after follow-up started (t>0).}

\item{xMat}{non-SNP covariates matrix.}

\item{gMat}{n*p genotype matrix.}

\item{coxph}{Boolean, whether to fit Cox PH model.}

\item{survreg}{Boolean, whether to fit survreg() weibull model.}

\item{obsInd}{n*1 vector of whether the event was observed or right-censored (0).}
}
\value{
A list with the elements:
\item{pVec}{p*1 vector of marginal SNP p-values}
}
\description{
Take a matrix of SNPs and get the interval-censored regression p-value for each one separately using either 
survreg() or coxph() with midpoint approximation.
}
\examples{
X <- matrix(data=rnorm(200), nrow=100)
lt <- runif(n=100, min=0, max=5)
rt <- lt + runif(n=100, min=0, max=5)
dmats <- make_IC_dmat(X=X, lt=lt, rt=rt)
null_fit <- skat_fit_null(init_beta=rep(0, 5), left_dmat=dmats$left_dmat,
right_dmat=dmats$right_dmat, obs_ind=rep(1, n), tpos_ind = as.numeric(lt > 0))
ICsingleSNP(left_dmat=dmats$left_dmat, right_dmat=dmats$right_dmat, obs_ind=rep(1, n),
tpos_ind = as.numeric(lt > 0), null_beta=null_fit$beta_fit, Itt=null_fit$Itt, 
gMat=matrix(data=rbinom(n=200*10, size=2, prob=0.3), nrow=200))

}
